Split interfacefiles (#423)
* Split interfaces into separate files

* Remove dependency of FluentExpandoObject on NetDaemonAppBase
