+++ src/App/NetDaemon.App/Common/FluentExpandoObject.cs
@@ -22,9 +22,9 @@ namespace NetDaemon.Common
     public class FluentExpandoObject : DynamicObject, IDictionary<string, object>
     {
         private readonly Dictionary<string, object> _dict = new();
-        private readonly NetDaemonAppBase? _daemonApp;
         private readonly bool _ignoreCase;
         private readonly bool _returnNullMissingProperties;
+        private readonly Action? _persistCallback;
 
         /// <summary>
         ///     Creates a BetterExpando object/
@@ -32,14 +32,14 @@ namespace NetDaemon.Common
         /// <param name="ignoreCase">Don't be strict about property name casing.</param>
         /// <param name="returnNullMissingProperties">If true, returns String.Empty for missing properties.</param>
         /// <param name="root">An ExpandoObject to consume and expose.</param>
-        /// <param name="daemon">A NetDaemon object used for persistanse</param>
+        /// <param name="persistCallback">callback method to persist on change</param>
         public FluentExpandoObject(bool ignoreCase = false,
             bool returnNullMissingProperties = false,
-            ExpandoObject? root = null, NetDaemonAppBase? daemon = null)
+            ExpandoObject? root = null, Action? persistCallback = null)
         {
-            _daemonApp = daemon;
             _ignoreCase = ignoreCase;
             _returnNullMissingProperties = returnNullMissingProperties;
+            _persistCallback = persistCallback;
             if (root != null) Augment(root);
         }
 
@@ -137,8 +137,8 @@ namespace NetDaemon.Common
                 throw new NetDaemonNullReferenceException(nameof(binder));
 
             UpdateDictionary(binder.Name, value);
-            // It is supposed to persist, this is the only reason _daemon is present
-            _daemonApp?.SaveAppState();
+            // It is supposed to persist
+            _persistCallback?.Invoke();
 
             return true;
         }
